Lower sounds must have a greater amplitude in order to have the same energy

+++ radio/src/audio_arm.h
@@ -106,7 +106,7 @@ class AudioContext {
       struct {
         double step;
         double idx;
-        double volume;
+        int volume;
       } tone;
     } state;
